requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
distlib==0.3.9
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
mysqlclient==2.2.7
oauthlib==3.3.1
platformdirs==4.3.7
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyMySQL==1.1.1
pyparsing==3.2.3
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.30.0
